Move config inside command.
