+++ hatch/commands/__init__.py
@@ -1,2 +1,3 @@
 from hatch.commands.conda import conda
 from hatch.commands.python import python
+from hatch.commands.config import config
